Hello To World
Hi, I'm Mojtaba, a new member of GitHub. I hope we can have the power to solve all the problems together. I love the irregular and disjointed login system very much and I have a high and terrible power in this field. Exactly when I look in front of me and my hand is pointing to the right and the knife is in my left hand, I attack from behind and I don't follow any rules for penetration at all. Self-made and brutal rules. This was the initial plan. Now I have a plan with which a more advanced replacement can be placed next to the global system.